requirements.txt
flask
selenium
numpy
Pillow
matplotlib
